#151 Added deckDao unit tests

+++ app/src/androidTest/java/org/liberty/android/fantastischmemo/test/domain/DeckTest.java
@@ -62,7 +62,7 @@ public class DeckTest {
     @Test
     public void testGetSetRating(){
         deck.setRating(3);
-        assertEquals(3,deck.getRating());
+        assertEquals((Integer)3,((Integer)deck.getRating()));
     }
